+++ src/components/Main.jsx
@@ -1,10 +1,12 @@
 import { Outlet } from "react-router-dom";
 import styles from "../styles/Main.module.css";
+import ScrollButtons from "./ScrollButtons";
 
 const Main = () => {
   return (
     <main className={styles.main}>
-      <Outlet />
+      <Outlet context={ScrollButtons} />
+      <ScrollButtons />
     </main>
   );
 };